+++ RELEASENOTES.md
@@ -1,6 +1,9 @@
 ## 0.9.1
 
 * adapt to runtime changes
+* do not emit MulleObjCLoader classes and categories when dumping dependencies
+* emit classid,categorid as comment when dumping dependencies
+
 
 ### 0.8.3